requirements.txt
htbuilder==0.6.1
matplotlib==3.5.1
numpy==1.24.2
numpy==1.22.3
pandas==1.4.4
Pillow==9.5.0
Pillow==9.5.0
plotly==5.11.0
pmdarima==2.0.3
pymongo==4.3.3
scipy==1.9.1
statsmodels==0.13.2
streamlit==1.22.0
yfinance==0.2.18
